+++ docs/source/python_api/mlflow.pyfunc.rst
@@ -5,3 +5,8 @@ mlflow.pyfunc
     :members:
     :undoc-members:
     :show-inheritance:
+
+
+.. automodule:: mlflow.pyfunc.model
+    :members:
+    :undoc-members:
