+++ mlflow/pyfunc/__init__.py
@@ -92,7 +92,7 @@ from copy import deepcopy
 from mlflow.tracking.fluent import active_run, log_artifacts
 from mlflow import tracking
 from mlflow.models import Model
-from mlflow.pyfunc.model import save_model, log_model, PythonModel, PythonModelContext 
+from mlflow.pyfunc.model import _save_model, PythonModel, PythonModelContext
 from mlflow.pyfunc.utils import _get_code_dirs,\
         _warn_potentially_incompatible_py_version_if_necessary
 from mlflow.utils import PYTHON_VERSION
@@ -143,6 +143,119 @@ def add_to_model(model, loader_module, data=None, code=None, env=None, **kwargs)
         parms[ENV] = env
     return model.add_flavor(FLAVOR_NAME, **parms)
 
+
+def save_model(path, model_class, artifacts, parameters, conda_env=None, code_paths=None,
+               mlflow_model=Model()):
+    """
+    :param path: The path to which to save the Python model.
+    :param model_class: A ``type`` object referring to a subclass of ``PythonModel``, or the
+                        fully-qualified name of such a subclass. ``model_class`` defines
+                        how the model is loaded and how it performs inference.
+    :param artifacts: A dictionary containing ``<name, artifact_uri>`` pairs. Remote artifact URIs
+                      will be resolved to absolute filesystem paths, producing a dictionary of
+                      ``<name, absolute_path>`` pairs. ``model_class`` can reference these resolved
+                      pairs as the ``artifacts`` property of the ``context`` attribute.
+                      For example, consider the following ``artifacts`` dictionary::
+
+                        {
+                            "my_file": "s3://my-bucket/path/to/my/file"
+                        }
+
+                      In this case, the ``"my_file"`` artifact will be downloaded from S3. The
+                      ``model_class`` can then refer to ``"my_file"`` as an absolute path via
+                      ``self.context.artifacts["my_file"]``.
+    :param parameters: A dictionary containing ``<name, python_object>`` pairs. ``python_object``
+                       may be any Python object that is serializable with CloudPickle.
+                       ``model_class`` can reference these resolved pairs as the ``parameters``
+                       property of the ``context`` attribute. For example, consider the following
+                       ``parameters`` dictionary::
+
+                         {
+                             "my_list": range(10)
+                         }
+
+                       The ``model_class`` can refer to the Python list named ``"my_list"`` as
+                       ``self.context.parameters["my_list"]``.
+    :param conda_env: Either a dictionary representation of a Conda environment or the path to a
+                      Conda environment yaml file. If provided, this decribes the environment
+                      this model should be run in. At minimum, it should specify the dependencies
+                      contained in ``mlflow.pyfunc.model.DEFAULT_CONDA_ENV``. If `None`, the default
+                      ``mlflow.pyfunc.model.DEFAULT_CONDA_ENV`` environment will be added to the
+                      model. The following is an *example* dictionary representation of a Conda
+                      environment::
+
+                        {
+                            'name': 'mlflow-env',
+                            'channels': ['defaults'],
+                            'dependencies': [
+                                'python=3.7.0',
+                                'cloudpickle=0.5.8'
+                            ]
+                        }
+    :param code_paths: A list of paths to Python file dependencies that are required by
+                       instances of ``model_class``.
+    :param mlflow_model: The model configuration to which to add the ``mlflow.pyfunc`` flavor.
+    """
+    return _save_model(path=path, model_class=model_class, artifacts=artifacts,
+                       parameters=parameters, conda_env=conda_env, code_paths=code_paths,
+                       mlflow_model=mlflow_model)
+
+
+def log_model(artifact_path, artifacts, parameters, model_class, conda_env=None,
+              code_paths=None):
+    """
+    :param path: The run-relative artifact path to which to log the Python model.
+    :param model_class: A ``type`` object referring to a subclass of ``PythonModel``, or the
+                        fully-qualified name of such a subclass. ``model_class`` defines
+                        how the model is loaded and how it performs inference.
+    :param artifacts: A dictionary containing ``<name, artifact_uri>`` pairs. Remote artifact URIs
+                      will be resolved to absolute filesystem paths, producing a dictionary of
+                      ``<name, absolute_path>`` pairs. ``model_class`` can reference these resolved
+                      pairs as the ``artifacts`` property of the ``context`` attribute.
+                      For example, consider the following ``artifacts`` dictionary::
+
+                        {
+                            "my_file": "s3://my-bucket/path/to/my/file"
+                        }
+
+                      In this case, the ``"my_file"`` artifact will be downloaded from S3. The
+                      ``model_class`` can then refer to ``"my_file"`` as an absolute path via
+                      ``self.context.artifacts["my_file"]``.
+    :param parameters: A dictionary containing ``<name, python_object>`` pairs. ``python_object``
+                       may be any Python object that is serializable with CloudPickle.
+                       ``model_class`` can reference these resolved pairs as the ``parameters``
+                       property of the ``context`` attribute. For example, consider the following
+                       ``parameters`` dictionary::
+
+                         {
+                             "my_list": range(10)
+                         }
+
+                       The ``model_class`` can refer to the Python list named ``"my_list"`` as
+                       ``self.context.parameters["my_list"]``.
+    :param conda_env: Either a dictionary representation of a Conda environment or the path to a
+                      Conda environment yaml file. If provided, this decribes the environment
+                      this model should be run in. At minimum, it should specify the dependencies
+                      contained in ``mlflow.pyfunc.model.DEFAULT_CONDA_ENV``. If `None`, the default
+                      ``mlflow.pyfunc.model.DEFAULT_CONDA_ENV`` environment will be added to the
+                      model. The following is an *example* dictionary representation of a Conda
+                      environment::
+
+                        {
+                            'name': 'mlflow-env',
+                            'channels': ['defaults'],
+                            'dependencies': [
+                                'python=3.7.0',
+                                'cloudpickle=0.5.8'
+                            ]
+                        }
+    :param code_paths: A list of paths to Python file dependencies that are required by
+                       instances of ``model_class``.
+    :param mlflow_model: The model configuration to which to add the ``mlflow.pyfunc`` flavor.
+    """
+    return Model.log(artifact_path=artifact_path, flavor=__name__, artifacts=artifacts,
+                     parameters=parameters, model_class=model_class, conda_env=conda_env,
+                     code_paths=code_paths)
 
 
 def load_pyfunc(path, run_id=None, suppress_warnings=False):
